Fix a spring loader error
When everything is in the same file, it's very easy to trigger the
following error:
"superclass mismatch for class RankingCondition"

This commits splits each class to its own file.
