+++ WcaOnRails/lib/advancement_conditions/ranking_condition.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module AdvancementConditions
+  class RankingCondition < AdvancementCondition
+    alias_method :ranking, :level
+
+    def self.wcif_type
+      "ranking"
+    end
+
+    def to_s(round, short: false)
+      I18n.t("advancement_condition#{".short" if short}.ranking", ranking: ranking)
+    end
+  end
+end
