+++ WcaOnRails/spec/models/competition_wcif_spec.rb
@@ -22,7 +22,7 @@ RSpec.describe "Competition WCIF" do
   }
   let(:delegate) { competition.delegates.first }
   let(:sixty_second_2_attempt_cutoff) { Cutoff.new(number_of_attempts: 2, attempt_result: 1.minute.in_centiseconds) }
-  let(:top_16_advance) { RankingCondition.new(16) }
+  let(:top_16_advance) { AdvancementConditions::RankingCondition.new(16) }
   let!(:round333_1) { FactoryBot.create(:round, competition: competition, event_id: "333", number: 1, cutoff: sixty_second_2_attempt_cutoff, advancement_condition: top_16_advance, scramble_set_count: 16, total_number_of_rounds: 2) }
   let!(:round333_2) { FactoryBot.create(:round, competition: competition, event_id: "333", number: 2, total_number_of_rounds: 2) }
   let!(:round444_1) { FactoryBot.create(:round, competition: competition, event_id: "444", number: 1) }
